prbench-rl/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "prbench_rl"
version = "0.1.0"
description = "Reinforcement learning baselines for PRBench."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
   "matplotlib",
   "numpy",
   "pandas",
   "pandas-stubs",
   "hydra-core",
   "omegaconf",
   "torch",
   "dacite",
   "gymnasium[classic-control]==0.29.1",
   "tensorboard",
]

[project.optional-dependencies]
develop = [
    "black",
    "docformatter",
    "isort",
    "mypy",
    "pylint>=2.14.5",
    "pytest-pylint>=0.18.0",
    "pytest>=7.2.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
prbench_bilevel_planning = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = "(third-party|\\.venv)"

[tool.isort]
py_version = 311
profile = "black"
multi_line_output = 2
skip_glob = ["venv/*", ".venv/*", "third-party/*"]
split_on_trailing_comma = true

[tool.docformatter]
line-length = 88
wrap-summaries = 88
wrap-descriptions = 88
exclude = ["venv", ".venv", "third-party"]

[tool.mypy]
strict_equality = true
disallow_untyped_calls = true
warn_unreachable = true
exclude = ["venv/*", ".venv/*", "third-party/*"]

[[tool.mypy.overrides]]
module = [
    "matplotlib.*",
    "torch.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--ignore=third-party"]
